Updated User schema

+++ api/build/controllers/vehicleController.js
@@ -72,10 +72,10 @@ const editVehicle = async (req, res, next) => {
 };
 exports.editVehicle = editVehicle;
 const deleteVehicle = async (req, res, next) => {
-    // let user: IUsers | null = await verifyUser(req);
-    // if (!user || user._id != req.body.Posted_By) {
-    //     return res.status(403).send();
-    // }
+    let user = await (0, auth_1.verifyUser)(req);
+    if (!user || user._id != req.body.Posted_By) {
+        return res.status(403).send();
+    }
     let itemId = req.params.id;
     let result = await Vehicle_1.Vehicles.findByIdAndDelete(itemId);
     res.status(200).json(result);
